Rename old TranslationView and clean it up
Rename the old TranslationView to InlineTranslationView (used in the
inline translation mode). Also remove all the unnecessary pieces.
